OCR/requirements.txt
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.24.0
customtkinter>=5.2.0
reportlab>=4.0.0
python-docx>=1.0.0
PyMuPDF>=1.24.0
